feat : create meeting index 페이지 구조 개선
